# schemas/web-api.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "JeRyu Web Forge API",
    "description": "BFF for the JeRyu Web Forge. Versioned under /api/v1/. See WEB_WORK_CLAUDE.md §35.7 for the canonical route map.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0-alpha"
  },
  "paths": {
...
    "/api/v1/repos": {
      "get": {
        "tags": [
          "repos"
        ],
        "operationId": "list_repos",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "host",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "owner",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "family",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "include_archived",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Repository list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepositoryListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "post": {
        "tags": [
          "repos"
        ],
        "operationId": "create_repo",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRepositoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Created repository",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepositorySummary"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "csrf": []
          }
        ]
      }
    },
...
    "/api/v1/repos/{repo_id}": {
      "get": {
        "tags": [
          "repos"
        ],
        "operationId": "get_repo",
        "parameters": [
          {
            "name": "repo_id",
            "in": "path",
            "description": "Stable opaque repo ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Repository detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepoDetailResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Repository not found"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "patch": {
        "tags": [
          "repos"
        ],
        "operationId": "patch_repo",
        "parameters": [
          {
            "name": "repo_id",
            "in": "path",
            "description": "Stable opaque repo ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Patched repository settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepositorySettings"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Repository not found"
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "csrf": []
          }
        ]
      }
    },
...
    "/api/v1/repos/{repo_id}/blame": {
      "get": {
        "tags": [
          "repo_browser"
        ],
        "operationId": "get_blame",
        "parameters": [
          {
            "name": "repo_id",
            "in": "path",
            "description": "Stable opaque repo ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ref",
            "in": "query",
            "description": "Branch/tag/SHA",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "path",
            "in": "query",
            "description": "File path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Per-line blame",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlameResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Path or ref not found"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/v1/repos/{repo_id}/blob": {
      "get": {
        "tags": [
          "repo_browser"
        ],
        "operationId": "get_blob",
        "parameters": [
          {
            "name": "repo_id",
            "in": "path",
            "description": "Stable opaque repo ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ref",
            "in": "query",
            "description": "Branch/tag/SHA",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "path",
            "in": "query",
            "description": "File path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "render",
            "in": "query",
            "description": "Optional renderer (e.g. `markdown`)",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Blob content + metadata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlobResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Path or ref not found"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
    "/api/v1/repos/{repo_id}/commits": {
      "get": {
        "tags": [
          "repo_browser"
        ],
        "operationId": "list_commits",
        "parameters": [
          {
            "name": "repo_id",
            "in": "path",
            "description": "Stable opaque repo ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ref",
            "in": "query",
            "description": "Branch/tag/SHA",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "path",
            "in": "query",
            "description": "Optional file path filter",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "description": "Page size (default 50)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Pagination cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Commits page",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommitsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Ref not found"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/v1/repos/{repo_id}/compare": {
      "get": {
        "tags": [
          "repo_browser"
        ],
        "operationId": "compare_refs",
        "parameters": [
          {
            "name": "repo_id",
            "in": "path",
            "description": "Stable opaque repo ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "base",
            "in": "query",
            "description": "Base ref",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "head",
            "in": "query",
            "description": "Head ref",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Diff comparison",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompareResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Ref not found"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
    "/api/v1/repos/{repo_id}/raw": {
      "get": {
        "tags": [
          "repo_browser"
        ],
        "operationId": "get_raw",
        "parameters": [
          {
            "name": "repo_id",
            "in": "path",
            "description": "Stable opaque repo ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ref",
            "in": "query",
            "description": "Branch/tag/SHA",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "path",
            "in": "query",
            "description": "File path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Raw file bytes; Content-Type reflects detected MIME"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Path or ref not found"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/v1/repos/{repo_id}/readme": {
      "get": {
        "tags": [
          "repo_browser"
        ],
        "operationId": "get_readme",
        "parameters": [
          {
            "name": "repo_id",
            "in": "path",
            "description": "Stable opaque repo ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ref",
            "in": "query",
            "description": "Branch/tag/SHA (default: HEAD)",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Rendered README",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlobResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "README not found"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/v1/repos/{repo_id}/refs": {
      "get": {
        "tags": [
          "repo_browser"
        ],
        "operationId": "list_refs",
        "parameters": [
          {
            "name": "repo_id",
            "in": "path",
            "description": "Stable opaque repo ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Branch/tag refs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Repository not found"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
    "/api/v1/repos/{repo_id}/tree": {
      "get": {
        "tags": [
          "repo_browser"
        ],
        "operationId": "get_tree",
        "parameters": [
          {
            "name": "repo_id",
            "in": "path",
            "description": "Stable opaque repo ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ref",
            "in": "query",
            "description": "Branch/tag/SHA",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "path",
            "in": "query",
            "description": "Path within the tree (default: repo root)",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tree entries at path",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TreeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Path or ref not found"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BlameResponse": {
        "type": "object",
        "required": [
          "lines"
        ],
        "properties": {
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BlameLine"
            }
          }
        }
      },
...
      "CommitsResponse": {
        "type": "object",
        "required": [
          "commits"
        ],
        "properties": {
          "commits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommitItem"
            }
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "RefsResponse": {
        "type": "object",
        "required": [
          "refs"
        ],
        "properties": {
          "refs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RefSelectorItem"
            }
          }
        }
      },
...
      "RepoDetailResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/RepositorySummary"
          }
        ]
      },
...
      "RepositoryListResponse": {
        "type": "object",
        "required": [
          "generated_at",
          "total",
          "repositories",
          "facets"
        ],
        "properties": {
          "facets": {
            "$ref": "#/components/schemas/RepositoryFacets"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time"
          },
          "repositories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RepositorySummary"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
    },
    "securitySchemes": {
      "csrf": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      },
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jeryu_session"
      }
    }
  },
...
}